Removed third party libraries from repo
These files are no longer needed in the repository, as they are now found via find_package in CMake. This requires that the user has them installed on their system and that CMake knows where to look for them, but using find_package is the proper way of referencing external libraries in CMake.